#56 플래너 토론장 다른 파일로 분리
